pyproject.toml
[tool.poetry]
name = "discordly"
version = "0.1.0"
description = ""
authors = ["Arpan Pandey <arpan@hackersreboot.tech>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.21.0"
"discord.py" = "^2.0.1"
requests = "^2.28.1"
appwrite = "^1.1.0"
fastapi = "^0.85.1"
uvicorn = "^0.18.3"
table2ascii = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
